Split weight transfer into WeightTransferClient
